Reorganized buffer file and added check for having loaded psd_length continuous background
